'Introduction' Changes:
Divided the introduction function to two different conditions: if the local user has invited the remote user, or vice versa. This solution makes it possible for both the client and server side to use the same function.

The main introduction function has also been divided to small functions that implement the steps of the whole introduction.